Adding DEV-Buttons to manually reset the databases & handleing startup
DB now gets copied from the assets folder to the app on first startup

Added DEV-Buttons where it is possible to either initilize the databases via the .csv files or to copy from the assets folder.

The .csv method might be used to update a outdated version in the assets folder manually for now. Otherwise it should not be used because it is very slow compared to the 2nd method